requirements.txt
requests==2.26.0
GitPython==3.1.24
groq